requirements.txt
fastapi~=0.115.12
uvicorn
sqlalchemy~=2.0.40
psycopg2
pydantic~=2.11.0
httpx~=0.28.1
python-dotenv~=1.1.0

PyJWT~=2.10.1
alembic~=1.15.2